+++ packages/vsf-storyblok-extension/fullSync.js
@@ -1,20 +1,9 @@
 import { syncStories } from './sync-stories'
-import { log } from './helpers'
+import { log, createIndex } from './helpers'
 
-export const fullSync = async (db, config, storyblokClient) => {
+export const fullSync = async (db, config) => {
   log('Syncing published stories!')
   await db.indices.delete({ ignore_unavailable: true, index: 'storyblok_stories' })
-  await db.indices.create({
-    index: 'storyblok_stories',
-    body: {
-      index: {
-        mapping: {
-          total_fields: {
-            limit: config.storyblok.fieldLimit || 1000
-          }
-        }
-      }
-    }
-  })
-  await syncStories({ db, perPage: config.storyblok.perPage, storyblokClient })
+  await db.indices.create(createIndex(config))
+  await syncStories({ db, perPage: config.storyblok.perPage })
 }
